requirements.txt
aiogram==3.12.0
asyncpg
aiohttp==3.9.3
python-dotenv==1.0.0